finition du site

+++ src/components/Admin.jsx
@@ -6,6 +6,13 @@ import { useNavigate } from "react-router-dom";
 import { auth } from '../firebaseConfig';
 import { Link } from 'react-router-dom';
 import { signOut } from 'firebase/auth'
+import image from "../images/Arc.png"
+import home from "../images/accueil.png"
+import edit from "../images/editer.png"
+import poubelle from "../images/poubelle1.png"
+import out from "../images/sortie.png"
+import add from "../images/ajouter.png"
+import { Modal, Button } from 'react-bootstrap';
 
 const Admin = () => {
 
@@ -20,7 +27,10 @@ const Admin = () => {
         });
       };
     const [postLists, setPostLists] = useState([]);
-    const postsCollectionRef = collection(db, "posts")
+    const postsCollectionRef = collection(db, "posts");
+    const [showModal, setShowModal] = useState(false); 
+  const [postToDelete, setPostToDelete] = useState(null); 
+
     let navigate = useNavigate();
  
     useEffect (() =>{
@@ -32,10 +42,27 @@ const Admin = () => {
      getPosts();
     },[])
  
-    const deletePost = async (id) =>{
-     const postDoc = doc(db, "posts", id)
-   await deleteDoc(postDoc)
+  const handleDelete = async () => {
+    if (postToDelete) {
+      const postDoc = doc(db, "posts", postToDelete);
+      await deleteDoc(postDoc);
+      setPostLists((posts) =>
+        posts.filter((post) => post.id !== postToDelete)
+      );
+      setShowModal(false); 
+      setPostToDelete(null);
     }
+  };
+
+  const openDeleteModal = (id) => {
+    setPostToDelete(id); 
+    setShowModal(true); 
+  };
+
+  const closeDeleteModal = () => {
+    setShowModal(false); 
+    setPostToDelete(null); 
+  };
  
     const handleEdit = (id) => {
      navigate(`/CreatePost/${id}`);
@@ -76,68 +103,145 @@ const Admin = () => {
    
     return ( 
 
-        <div className="HomeContainer"> 
+        <div className="HomeContainer" style={{display:'flex', flexDirection:'row', height:'100%'}}> 
+        <div
+      className="d-flex flex-column flex-shrink-0 bg-body-tertiary"
+      style={{ width: '4.5rem' }}
+    >
+      <a
+        href="/"
+        className="d-block p-3 link-body-emphasis text-decoration-none"
+        title="Icon-only"
+        data-bs-toggle="tooltip"
+        data-bs-placement="right"
+      >
+        <svg className="bi pe-none" width="40" height="32">
+          <use xlinkHref="#bootstrap" />
+        </svg>
+        <span className="visually-hidden">Icon-only</span>
+      </a>
 
+      <ul className="nav nav-pills nav-flush flex-column mb-auto text-center">
+        <li className="nav-item">
+          <Link
+            to='/admin'
+            className="nav-link active py-3 border-bottom rounded-0"
+            aria-current="page"
+            title="Home"
+            data-bs-toggle="tooltip"
+            data-bs-placement="right"
+          >
+            <div style={{width:'25px', height:'25px'}}>
+              <img style={{width:'100%', height:'100%'}} src={home}/>
+            </div>
+          </Link>
+        </li>
+        <li>
+          <Link
+            to="/CreatePost"
+            className="nav-link py-3 border-bottom rounded-0"
+            title="Dashboard"
+            data-bs-toggle="tooltip"
+            data-bs-placement="right"
+          >
+            <div style={{width:'25px', height:'25px'}}>
+              <img style={{width:'100%', height:'100%'}} src={add}/>
+            </div>
+          </Link>
+        </li>
+        <li>
+          <Link
+            to=""
+            className="nav-link py-3 border-bottom rounded-0"
+            title="Dashboard"
+            data-bs-toggle="tooltip"
+            data-bs-placement="right"
+          >
+            <div style={{width:'25px', height:'25px'}}>
+              <img style={{width:'100%', height:'100%'}} src={edit}/>
+            </div>
+          </Link>
+        </li>
+        <li>
+          <Link
+            className="nav-link py-3 border-bottom rounded-0"
+            title="Orders"
+            data-bs-toggle="tooltip"
+            data-bs-placement="right"
+            onClick={signUserOut}
+          >
+            <div style={{width:'25px', height:'25px'}} >
+               <img style={{width:'100%', height:'100%'}} src={out}/>  
+              </div>
+          </Link> 
+        </li>
+      </ul>
 
-        <div>
-        <button style={{width:"75px", height:"75px"}} onClick={signUserOut}> Log out </button>
-          <div>
-                  <div className="btn-group me-2">
-            <button style={{width:"75px", height:"75px"}} onClick={handleClick} type="button" className="btn btn-sm btn-outline-secondary">Add new post</button>
-          </div>
+    </div>
 
-        <div>
-          {postLists.map((post)=>{
-            return (
-            
-            <div key={post.id}> 
+        <div style={{width:'100%', height:'100vh', padding:'1%'}}>
+        
+          <div> 
 
-<div className="table-responsive small">
+          <button onClick={handleClick} type="button" className="btn btn-primary"> Nouveau </button>
+        <div style={{marginTop:'50px'}} className="table-responsive small">
         <table className="table table-striped table-sm">
-          <thead>
-            <tr>
-              <th scope="col">Title</th>
-              <th scope="col">Content</th>
-              <th scope="col">Start Date</th>
-              <th scope="col">End Date</th>
-              <th scope="col">Image/Video URL</th>
-              <th scope="col">Extra URL</th>
-              <th scope="col">Action</th>
-              
-            </tr>
-          </thead>
-          <tbody>
-            <tr>
-              <td>{truncateText(post.title, 11) || "vide"}</td>
-              {/* <td>{post.title}</td> */}
-              <td>{truncateText(post.content, 11) || "vide"}</td>
-              {/* <td>{post.content}</td> */}
-              <td>{post.debutdate || "vide"}</td>
-              <td>{post.enddate || "vide"}</td>
-              <td>{truncateText(post.mediaURL, 11) || "vide"}</td>
-              {/* <td>{post.mediaURL}</td> */}
-              <td>{truncateText(post.videoURL, 11) || "vide"}</td>
-              {/* <td>{post.videoURL || "N/A"}</td> */}
-              <td> {isAuth && (<button onClick={() => {
-            deletePost(post.id)
-            }}> 
-            &#128465; 
-          </button>)}
-          {isAuth && (<button onClick={() => {handleEdit(post.id)}}> edit </button>)}
-           </td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
+    <thead>
+      <tr>
+        <th scope="col">Title</th>
+        <th scope="col">Content</th>
+        <th scope="col">Image/Video URL</th>
+        <th scope="col">Extra URL</th>
+        <th scope="col">Action</th>
+      </tr>
+    </thead>
+    <tbody>
+      {postLists.map((post) => (
+        <tr key={post.id}>
+          <td>{truncateText(post.title, 11) || 'vide'}</td>
+          <td>{truncateText(post.content, 11) || 'vide'}</td>
+          <td>{truncateText(post.mediaURL, 11) || 'vide'}</td>
+          <td>{truncateText(post.videoURL, 11) || 'vide'}</td>
+          <td>
+          <button
+                      style={{ border: "none", width: "40px", height: "40px" }}
+                      onClick={() => openDeleteModal(post.id)}
+                    >
+                      <img style={{ width: "25px", height: "25px" }} src={poubelle} />
+                    </button>
+            <button style={{border:'none', width:'40px', height:'40px'}} onClick={() => handleEdit(post.id)}>
+              <img style={{width:'25px', height:'25px'}} src={edit}/>
+            </button>
+          </td>
+        </tr>
+      ))}
+    </tbody>
+  </table>
+
 
             </div>
-            )
-          })}
-        </div>
+
+            <Modal show={showModal} onHide={closeDeleteModal} centered>
+          <Modal.Header closeButton>
+            <Modal.Title>Confirmer la suppression</Modal.Title>
+          </Modal.Header>
+          <Modal.Body>
+            Êtes-vous sûr de vouloir supprimer cet élément ? Cette action est irréversible.
+          </Modal.Body>
+          <Modal.Footer>
+            <Button variant="secondary" onClick={closeDeleteModal}>
+              Annuler
+            </Button>
+            <Button variant="danger" onClick={handleDelete}>
+              Supprimer
+            </Button>
+          </Modal.Footer>
+        </Modal>
+            </div>
           
           </div>
         </div>
-      </div>
+     
      
      );
 }
